backend/requirements.txt
# FastAPI and async
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Async queue and workers
aioredis==2.0.1
celery==5.3.6
httpx==0.26.0

# Compression
zstandard==0.22.0

# Utils
python-dotenv==1.0.0
aiofiles==23.2.1
python-dateutil==2.8.2

# WebSocket
websockets==12.0

# GitHub API
pygithub==2.2.0

# Logging
loguru==0.7.2

# Optional: HuggingFace dataset upload
# huggingface-hub==0.20.0